Rename check_db to analysis_db
