+++ src/vm/analysis/analysis_db.rs
